requirements.txt
Django==4.2 
djangorestframework==3.14 
psycopg2-binary==2.9.6 
djangorestframework-simplejwt==5.1 
gunicorn==23.0.0